requirements.txt
streamlit
pypdf
langchain
langchain-openai
faiss-cpu
python-dotenv
tiktoken
